+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 <!-- Add new, unreleased items here. -->
+* Updated wct-browser-legacy to use a module version of a11ySuite to get access to Polymer.dom.flush.
+* Updated generated index for webserver to use a11ySuite as a module.
 
 ## 6.2.0 - 2017-09-19